requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.7
django-filter==2.0.0
django-picklefield==1.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
whitenoise==3.3.1